Create Merkle_Tree_Hash Generator
A python program to generate the Merkle tree hash of a random number of given hashes.
In the program, the number of hashes is set to 13. But that can be changed or the number of hashes can be taken as an input from the user as well.
Hope you like it.
Do point out any mistakes that you find.
I welcome any suggestions as well.
Thankyou.